requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Markdown==3.4.4
Flask-Login==0.6.2
python-slugify==8.0.1
